NIFI-10656 eliminate duplicate logging, creating separate logs for known and unknown errors, extract unit tests to separate test
