+++ nifi-nar-bundles/nifi-azure-bundle/nifi-azure-processors/src/test/java/org/apache/nifi/processors/azure/storage/ITPutAzureDataLakeStorage.java
@@ -16,24 +16,16 @@
  */
 package org.apache.nifi.processors.azure.storage;
 
-import com.azure.core.http.rest.Response;
 import com.azure.storage.file.datalake.DataLakeDirectoryClient;
 import com.azure.storage.file.datalake.DataLakeFileClient;
-import com.azure.storage.file.datalake.models.DataLakeRequestConditions;
-import com.azure.storage.file.datalake.models.DataLakeStorageException;
 import org.apache.commons.lang3.StringUtils;
-import org.apache.nifi.flowfile.FlowFile;
-import org.apache.nifi.processor.ProcessSession;
 import org.apache.nifi.processor.Processor;
-import org.apache.nifi.processor.ProcessorInitializationContext;
 import org.apache.nifi.provenance.ProvenanceEventRecord;
 import org.apache.nifi.provenance.ProvenanceEventType;
-import org.apache.nifi.util.MockComponentLog;
 import org.apache.nifi.util.MockFlowFile;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import java.io.InputStream;
 import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.HashMap;
@@ -42,22 +34,12 @@ import java.util.Random;
 import java.util.Set;
 import java.util.stream.Collectors;
 
-import static org.apache.nifi.processors.azure.storage.PutAzureDataLakeStorage.FAIL_RESOLUTION;
 import static org.apache.nifi.processors.azure.storage.utils.ADLSAttributes.ATTR_NAME_DIRECTORY;
 import static org.apache.nifi.processors.azure.storage.utils.ADLSAttributes.ATTR_NAME_FILENAME;
 import static org.apache.nifi.processors.azure.storage.utils.ADLSAttributes.ATTR_NAME_FILESYSTEM;
 import static org.apache.nifi.processors.azure.storage.utils.ADLSAttributes.ATTR_NAME_LENGTH;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.ArgumentMatchers.anyLong;
-import static org.mockito.ArgumentMatchers.anyString;
-import static org.mockito.ArgumentMatchers.isNull;
-import static org.mockito.Mockito.doThrow;
-import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.verify;
-import static org.mockito.Mockito.when;
 
 public class ITPutAzureDataLakeStorage extends AbstractAzureDataLakeStorageIT {
 
@@ -262,42 +244,6 @@ public class ITPutAzureDataLakeStorage extends AbstractAzureDataLakeStorageIT {
         runProcessor(FILE_DATA, attributes);
 
         assertFailure();
-    }
-
-    @Test
-    public void testPutFileButFailedToAppend() {
-        final PutAzureDataLakeStorage processor = new PutAzureDataLakeStorage();
-        final DataLakeFileClient fileClient = mock(DataLakeFileClient.class);
-        final ProcessSession session = mock(ProcessSession.class);
-        final FlowFile flowFile = mock(FlowFile.class);
-
-        when(flowFile.getSize()).thenReturn(1L);
-        doThrow(IllegalArgumentException.class).when(fileClient).append(any(InputStream.class), anyLong(), anyLong());
-
-        assertThrows(IllegalArgumentException.class, () -> processor.appendContent(flowFile, fileClient, session));
-        verify(fileClient).delete();
-    }
-
-    @Test
-    public void testPutFileButFailedToRename() {
-        final PutAzureDataLakeStorage processor = new PutAzureDataLakeStorage();
-        final ProcessorInitializationContext initContext = mock(ProcessorInitializationContext.class);
-        final String componentId = "componentId";
-        final DataLakeFileClient fileClient = mock(DataLakeFileClient.class);
-        final Response<DataLakeFileClient> response = mock(Response.class);
-        final DataLakeStorageException exception = mock(DataLakeStorageException.class);
-        //Mock logger
-        when(initContext.getIdentifier()).thenReturn(componentId);
-        MockComponentLog componentLog = new MockComponentLog(componentId, processor);
-        when(initContext.getLogger()).thenReturn(componentLog);
-        processor.initialize(initContext);
-        //Mock renameWithResponse Azure method
-        when(fileClient.renameWithResponse(isNull(), anyString(), isNull(), any(DataLakeRequestConditions.class), isNull(), isNull())).thenReturn(response);
-        when(fileClient.getFileName()).thenReturn(FILE_NAME);
-        when(exception.getStatusCode()).thenReturn(405);
-        when(response.getValue()).thenThrow(exception);
-        assertThrows(DataLakeStorageException.class, () -> processor.renameFile(fileClient, "", FILE_NAME, FAIL_RESOLUTION));
-        verify(fileClient).delete();
     }
 
     private Map<String, String> createAttributesMap() {
